smol: Refactor module into separate files
